Rename MakeBrighterFatterKernelTaskRunner and move to utils
Other parts of cp_pipe need paired visits, so move this
to a utils file and give a generic name, PairedVisitTaskRunner
