+++ python/lsst/cp/pipe/makeBrighterFatterKernel.py
@@ -40,6 +40,7 @@ import lsst.afw.display as afwDisp
 from lsst.ip.isr import IsrTask
 import lsst.pex.config as pexConfig
 import lsst.pipe.base as pipeBase
+from .utils import PairedVisitTaskRunner
 
 
 class MakeBrighterFatterKernelTaskConfig(pexConfig.Config):
@@ -166,40 +167,6 @@ class MakeBrighterFatterKernelTaskConfig(pexConfig.Config):
     )
 
 
-class MakeBrighterFatterKernelTaskRunner(pipeBase.TaskRunner):
-    """Subclass of TaskRunner for the MakeBrighterFatterKernelTask.
-
-    This transforms the processed arguments generated by the ArgumentParser
-    into the arguments expected by makeBrighterFatterKernelTask.run().
-
-    makeBrighterFatterKernelTask.run() takes a two arguments,
-    one of which is the dataRef (as usual), and the other is the list
-    of visit-pairs, in the form of a list of tuples.
-    This list is supplied on the command line as documented,
-    and this class parses that, and passes the parsed version
-    to the run() method.
-
-    See pipeBase.TaskRunner for more information.
-    """
-
-    @staticmethod
-    def getTargetList(parsedCmd, **kwargs):
-        """Parse the visit list and pass through explicitly."""
-        visitPairs = []
-        for visitStringPair in parsedCmd.visitPairs:
-            visitStrings = visitStringPair.split(",")
-            if len(visitStrings) != 2:
-                raise RuntimeError("Found {} visits in {} instead of 2".format(len(visitStrings),
-                                                                               visitStringPair))
-            try:
-                visits = [int(visit) for visit in visitStrings]
-            except Exception:
-                raise RuntimeError("Could not parse {} as two integer visit numbers".format(visitStringPair))
-            visitPairs.append(visits)
-
-        return pipeBase.TaskRunner.getTargetList(parsedCmd, visitPairs=visitPairs, **kwargs)
-
-
 class BrighterFatterKernelTaskDataIdContainer(pipeBase.DataIdContainer):
     """A DataIdContainer for the MakeBrighterFatterKernelTask."""
 
@@ -306,7 +273,7 @@ class MakeBrighterFatterKernelTask(pipeBase.CmdLineTask):
       TODO: DM-15277 update this part of the docstring once the ticket is done.
     """
 
-    RunnerClass = MakeBrighterFatterKernelTaskRunner
+    RunnerClass = PairedVisitTaskRunner
     ConfigClass = MakeBrighterFatterKernelTaskConfig
     _DefaultName = "makeBrighterFatterKernel"
