+++ python/lsst/cp/pipe/utils.py
@@ -1,0 +1,60 @@
+# This file is part of cp_pipe.
+#
+# Developed for the LSST Data Management System.
+# This product includes software developed by the LSST Project
+# (https://www.lsst.org).
+# See the COPYRIGHT file at the top-level directory of this distribution
+# for details of code ownership.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+
+__all__ = ['PairedVisitTaskRunner', ]
+
+import lsst.pipe.base as pipeBase
+
+
+class PairedVisitTaskRunner(pipeBase.TaskRunner):
+    """Subclass of TaskRunner for handling intrinsically paired visits.
+
+    This transforms the processed arguments generated by the ArgumentParser
+    into the arguments expected by tasks which take visit pairs for their
+    run() methods.
+
+    Such tasks' run() methods tend to take two arguments,
+    one of which is the dataRef (as usual), and the other is the list
+    of visit-pairs, in the form of a list of tuples.
+    This list is supplied on the command line as documented,
+    and this class parses that, and passes the parsed version
+    to the run() method.
+
+    See pipeBase.TaskRunner for more information.
+    """
+
+    @staticmethod
+    def getTargetList(parsedCmd, **kwargs):
+        """Parse the visit list and pass through explicitly."""
+        visitPairs = []
+        for visitStringPair in parsedCmd.visitPairs:
+            visitStrings = visitStringPair.split(",")
+            if len(visitStrings) != 2:
+                raise RuntimeError("Found {} visits in {} instead of 2".format(len(visitStrings),
+                                                                               visitStringPair))
+            try:
+                visits = [int(visit) for visit in visitStrings]
+            except Exception:
+                raise RuntimeError("Could not parse {} as two integer visit numbers".format(visitStringPair))
+            visitPairs.append(visits)
+
+        return pipeBase.TaskRunner.getTargetList(parsedCmd, visitPairs=visitPairs, **kwargs)
